+++ app/src/main/java/org/linphone/activities/main/settings/viewmodels/CallSettingsViewModel.kt
@@ -290,7 +290,11 @@ class CallSettingsViewModel : GenericSettingsViewModel() {
             encryptionValues.add(MediaEncryption.SRTP.toInt())
         }
         if (core.mediaEncryptionSupported(MediaEncryption.ZRTP)) {
-            labels.add(prefs.getString(R.string.call_settings_media_encryption_zrtp))
+            if (core.postQuantumAvailable) {
+                labels.add(prefs.getString(R.string.call_settings_media_encryption_zrtp_post_quantum))
+            } else {
+                labels.add(prefs.getString(R.string.call_settings_media_encryption_zrtp))
+            }
             encryptionValues.add(MediaEncryption.ZRTP.toInt())
         }
         if (core.mediaEncryptionSupported(MediaEncryption.DTLS)) {